+++ src/app/pages/Profile/MyGallery/MyGallery.module.scss
